+++ backend/src/classes/Brave/Core/DataFixtures/RoleFixtureLoader.php
@@ -1,0 +1,44 @@
+<?php declare(strict_types=1);
+
+namespace Brave\Core\DataFixtures;
+
+use Brave\Core\Entity\Role;
+use Brave\Core\Factory\RepositoryFactory;
+use Doctrine\Common\DataFixtures\FixtureInterface;
+use Doctrine\Common\Persistence\ObjectManager;
+
+class RoleFixtureLoader implements FixtureInterface
+{
+    public function load(ObjectManager $manager)
+    {
+        $roleRepository = (new RepositoryFactory($manager))->getRoleRepository();
+
+        $roles = [
+            1 => Role::USER,
+            2 => Role::APP,
+            3 => Role::USER_ADMIN,
+            4 => Role::GROUP_ADMIN,
+            5 => Role::GROUP_MANAGER,
+            6 => Role::APP_ADMIN,
+            7 => Role::APP_MANAGER,
+            8 => Role::ESI,
+            9 => Role::SETTINGS,
+            10 => Role::TRACKING,
+            11 => Role::APP_TRACKING,
+            12 => Role::APP_ESI,
+            13 => Role::APP_GROUPS,
+            14 => Role::APP_CHARS,
+        ];
+
+        foreach ($roles as $id => $name) {
+            $role = $roleRepository->find($id);
+            if ($role === null) {
+                $role = new Role($id);
+                $manager->persist($role);
+            }
+            $role->setName($name);
+        }
+
+        $manager->flush();
+    }
+}