+++ backend/src/classes/Brave/Core/Migrations/Version20180410183906.php
@@ -2,7 +2,6 @@
 
 namespace Brave\Core\Migrations;
 
-use Brave\Core\Entity\Role;
 use Doctrine\Migrations\AbstractMigration;
 use Doctrine\DBAL\Schema\Schema;
 
@@ -13,11 +12,7 @@ class Version20180410183906 extends AbstractMigration
 {
     public function up(Schema $schema) : void
     {
-        $this->addSql('INSERT INTO roles (id, name) VALUES (3, "'.Role::USER_ADMIN.'")');
-        $this->addSql('INSERT INTO roles (id, name) VALUES (4, "'.Role::GROUP_ADMIN.'")');
-        $this->addSql('INSERT INTO roles (id, name) VALUES (5, "'.Role::GROUP_MANAGER.'")');
-        $this->addSql('INSERT INTO roles (id, name) VALUES (6, "'.Role::APP_ADMIN.'")');
-        $this->addSql('INSERT INTO roles (id, name) VALUES (7, "'.Role::APP_MANAGER.'")');
+        // role(s) were inserted here
     }
 
     public function down(Schema $schema) : void