+++ backend/src/classes/Brave/Core/Migrations/Version20180416212822.php
@@ -1,6 +1,7 @@
 <?php
 
 declare(strict_types = 1);
+
 namespace Brave\Core\Migrations;
 
 use Doctrine\Migrations\AbstractMigration;
@@ -13,6 +14,7 @@ class Version20180416212822 extends AbstractMigration
 {
     public function up(Schema $schema) : void
     {
+        // Previously an app was added in Version20171229114138
         $this->addSql('DELETE FROM apps WHERE id = :id', array('id' => 1));
     }