+++ backend/src/classes/Brave/Core/Migrations/Version20180806193042.php
@@ -2,7 +2,6 @@
 
 namespace Brave\Core\Migrations;
 
-use Brave\Core\Entity\Role;
 use Doctrine\DBAL\Schema\Schema;
 use Doctrine\Migrations\AbstractMigration;
 
@@ -13,9 +12,7 @@ final class Version20180806193042 extends AbstractMigration
 {
     public function up(Schema $schema) : void
     {
-        // this up() migration is auto-generated, please modify it to your needs
-
-        $this->addSql('INSERT INTO roles (id, name) VALUES (8, "'.Role::ESI.'")');
+        // role(s) were inserted here
     }
 
     public function down(Schema $schema) : void