+++ backend/src/classes/Brave/Core/Migrations/Version20181120205730.php
@@ -2,7 +2,6 @@
 
 namespace Brave\Core\Migrations;
 
-use Brave\Core\Entity\SystemVariable;
 use Doctrine\DBAL\Schema\Schema;
 use Doctrine\Migrations\AbstractMigration;
 
@@ -11,25 +10,12 @@ use Doctrine\Migrations\AbstractMigration;
  */
 final class Version20181120205730 extends AbstractMigration
 {
-    /**
-     * @throws \Exception
-     */
     public function up(Schema $schema) : void
     {
-        // this up() migration is auto-generated, please modify it to your needs
-        $this->abortIf($this->connection->getDatabasePlatform()->getName() !== 'mysql', 'Migration can only be executed safely on \'mysql\'.');
-
-        $this->addSql('INSERT INTO system_variables (name, value, scope) VALUES ("'.SystemVariable::MAIL_ACCOUNT_DISABLED_ALLIANCES.'", "", "settings")');
+        // setting variable(s) were inserted here
     }
 
-    /**
-     * @throws \Exception
-     */
     public function down(Schema $schema) : void
     {
-        // this down() migration is auto-generated, please modify it to your needs
-        $this->abortIf($this->connection->getDatabasePlatform()->getName() !== 'mysql', 'Migration can only be executed safely on \'mysql\'.');
-
-        $this->addSql('DELETE FROM system_variables WHERE name = "'.SystemVariable::MAIL_ACCOUNT_DISABLED_ALLIANCES.'"');
     }
 }