+++ backend/src/classes/Brave/Core/Migrations/Version20190210120518.php
@@ -2,7 +2,6 @@
 
 namespace Brave\Core\Migrations;
 
-use Brave\Core\Entity\Role;
 use Doctrine\DBAL\Schema\Schema;
 use Doctrine\Migrations\AbstractMigration;
 
@@ -11,27 +10,12 @@ use Doctrine\Migrations\AbstractMigration;
  */
 final class Version20190210120518 extends AbstractMigration
 {
-    /**
-     * @throws \Exception
-     */
     public function up(Schema $schema) : void
     {
-        // this up() migration is auto-generated, please modify it to your needs
-        $this->abortIf($this->connection->getDatabasePlatform()->getName() !== 'mysql',
-            'Migration can only be executed safely on \'mysql\'.');
-
-        $this->addSql('INSERT INTO roles (id, name) VALUES (12, "'.Role::APP_ESI.'")');
+        // role(s) were inserted here
     }
 
-    /**
-     * @throws \Exception
-     */
     public function down(Schema $schema) : void
     {
-        // this down() migration is auto-generated, please modify it to your needs
-        $this->abortIf($this->connection->getDatabasePlatform()->getName() !== 'mysql',
-            'Migration can only be executed safely on \'mysql\'.');
-
-        $this->addSql('DELETE FROM roles WHERE id = 12');
     }
 }