+++ backend/src/classes/Brave/Core/Migrations/Version20190310165548.php
@@ -4,7 +4,6 @@ declare(strict_types=1);
 
 namespace Brave\Core\Migrations;
 
-use Brave\Core\Entity\SystemVariable;
 use Doctrine\DBAL\Schema\Schema;
 use Doctrine\Migrations\AbstractMigration;
 
@@ -18,25 +17,12 @@ final class Version20190310165548 extends AbstractMigration
         return '';
     }
 
-    /**
-     * @throws \Doctrine\DBAL\DBALException
-     */
     public function up(Schema $schema) : void
     {
-        // this up() migration is auto-generated, please modify it to your needs
-        $this->abortIf($this->connection->getDatabasePlatform()->getName() !== 'mysql', 'Migration can only be executed safely on \'mysql\'.');
-
-        $this->addSql("INSERT INTO system_variables (name, value, scope) VALUES ('".SystemVariable::ALLOW_LOGIN_MANAGED."', '0', 'settings')");
+        // setting variable(s) were inserted here
     }
 
-    /**
-     * @throws \Doctrine\DBAL\DBALException
-     */
     public function down(Schema $schema) : void
     {
-        // this down() migration is auto-generated, please modify it to your needs
-        $this->abortIf($this->connection->getDatabasePlatform()->getName() !== 'mysql', 'Migration can only be executed safely on \'mysql\'.');
-
-        $this->addSql("DELETE FROM system_variables WHERE name = '".SystemVariable::ALLOW_LOGIN_MANAGED."'");
     }
 }